requirements.txt
Flask>=3.0.0
pyserial>=3.5
RPLCD>=1.3.1
smbus2>=0.4.3
RPi.GPIO>=0.7.1
pytest>=7.0.0
python-dotenv>=1.0.1
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
WTForms>=3.1.1
Werkzeug>=3.0.1

